fix: move covertToBool to helper
