+++ plugins/helper.js
@@ -364,6 +364,24 @@ async function processHooks(apiUri, token, webhookConfig) {
         });
 }
 
+/**
+ * convert value to Boolean
+ * @method convertToBool
+ * @param {(Boolean|String)} value
+ * @return {Boolean}
+ */
+function convertToBool(value) {
+    if (typeof value === 'boolean') {
+        return value;
+    }
+
+    // trueList refers to https://yaml.org/type/bool.html
+    const trueList = ['on', 'true', 'yes', 'y'];
+    const lowerValue = String(value).toLowerCase();
+
+    return trueList.includes(lowerValue);
+}
+
 module.exports = {
     requestRetryStrategy,
     requestRetryStrategyPostEvent,
@@ -376,5 +394,6 @@ module.exports = {
     getPipelineConfig,
     updateBuild,
     notifyJob,
-    processHooks
+    processHooks,
+    convertToBool
 };
