+++ plugins/queue/scheduler.js
@@ -11,7 +11,7 @@ const helper = require('../helper');
 const { timeOutOfWindows } = require('./utils/freezeWindows');
 const { queueNamespace } = require('../../config/redis');
 const ecosystem = Config.get('ecosystem');
-const { periodicBuildTableEnabled } = Config.get('queue');
+const periodicBuildTableEnabled = helper.convertToBool(Config.get('queue').periodicBuildTableEnabled);
 const DEFAULT_BUILD_TIMEOUT = 90;
 const RETRY_LIMIT = 3;
 const RETRY_DELAY = 5;
